move flag-related extensions from project-base to the framework package
- removed useless FlagFacadeTest
