+++ packages/framework/src/Form/Admin/Product/Flag/FlagFormType.php
@@ -4,11 +4,15 @@ declare(strict_types=1);
 
 namespace Shopsys\FrameworkBundle\Form\Admin\Product\Flag;
 
+use Shopsys\FormTypesBundle\YesNoType;
 use Shopsys\FrameworkBundle\Form\ColorPickerType;
-use Shopsys\FrameworkBundle\Form\Locale\LocalizedType;
+use Shopsys\FrameworkBundle\Form\GroupType;
+use Shopsys\FrameworkBundle\Form\LocalizedFullWidthType;
+use Shopsys\FrameworkBundle\Form\UrlListType;
+use Shopsys\FrameworkBundle\Model\Product\Flag\Flag;
 use Shopsys\FrameworkBundle\Model\Product\Flag\FlagData;
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Validator\Constraints;
@@ -22,7 +26,8 @@ class FlagFormType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('name', LocalizedType::class, [
+            ->add('name', LocalizedFullWidthType::class, [
+                'label' => t('Name'),
                 'required' => true,
                 'entry_options' => [
                     'constraints' => [
@@ -32,17 +37,47 @@ class FlagFormType extends AbstractType
                         ),
                     ],
                 ],
-            ])
+                'render_form_row' => false,
+            ]);
+
+        $builderBasicInformationGroup = $builder->create('basicInformationGroup', GroupType::class, [
+            'label' => t('Basic information'),
+        ]);
+
+        $builderBasicInformationGroup
             ->add('rgbColor', ColorPickerType::class, [
+                'label' => t('Color'),
                 'constraints' => [
                     new Constraints\NotBlank(['message' => 'Please enter flag color']),
                     new Constraints\Length([
                         'max' => 7,
-                        'maxMessage' => 'Flag color in must be in valid hexadecimal code e.g. #3333ff',
+                        'maxMessage' => 'Flag color must be in valid hexadecimal code e.g. #3333ff',
                     ]),
                 ],
             ])
-            ->add('visible', CheckboxType::class, ['required' => false]);
+            ->add('visible', YesNoType::class, [
+                'required' => false,
+                'label' => t('Display'),
+            ]);
+
+        $builder->add($builderBasicInformationGroup);
+
+        if ($options['flag'] !== null) {
+            $builderSeoInformationGroup = $builder->create('seoGroup', GroupType::class, [
+                'label' => t('Seo'),
+            ]);
+
+            $builderSeoInformationGroup
+                ->add('urls', UrlListType::class, [
+                    'route_name' => 'front_flag_detail',
+                    'entity_id' => $options['flag']->getId(),
+                    'label' => t('URL addresses'),
+                ]);
+
+            $builder->add($builderSeoInformationGroup);
+        }
+
+        $builder->add('save', SubmitType::class);
     }
 
     /**
@@ -50,9 +85,12 @@ class FlagFormType extends AbstractType
      */
     public function configureOptions(OptionsResolver $resolver): void
     {
-        $resolver->setDefaults([
-            'data_class' => FlagData::class,
-            'attr' => ['novalidate' => 'novalidate'],
-        ]);
+        $resolver
+            ->setRequired(['flag'])
+            ->setAllowedTypes('flag', [Flag::class, 'null'])
+            ->setDefaults([
+                'data_class' => FlagData::class,
+                'attr' => ['novalidate' => 'novalidate'],
+            ]);
     }
 }
