+++ packages/framework/src/Model/Product/Flag/FlagData.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Shopsys\FrameworkBundle\Model\Product\Flag;
 
+use Shopsys\FrameworkBundle\Component\Router\FriendlyUrl\UrlListData;
+
 class FlagData
 {
     /**
@@ -26,10 +28,16 @@ class FlagData
      */
     public $uuid;
 
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\Router\FriendlyUrl\UrlListData
+     */
+    public $urls;
+
     public function __construct()
     {
         $this->name = [];
         $this->visible = false;
         $this->rgbColor = '';
+        $this->urls = new UrlListData();
     }
 }
