+++ packages/framework/src/Resources/translations/messages.cs.po
@@ -1636,6 +1636,12 @@ msgstr "Editace e-mailové šablony - %name%"
 msgid "Editing file"
 msgstr "Upravit soubor"
 
+msgid "Editing flag"
+msgstr "Editace příznaku"
+
+msgid "Editing flag - {{ name }}"
+msgstr "Editace příznaku - {{ name }}"
+
 msgid "Editing holiday / internal day - {{ name }}"
 msgstr "Editace svátku / interního dne - {{ name }}"
 
@@ -1984,9 +1990,6 @@ msgstr "Vyplňte pouze URL pro zvolenou doménu"
 msgid "Fill missing settings to enable creating products."
 msgstr "Pro vytvoření produktu doplňte chybějící nastavení."
 
-msgid "Filter by"
-msgstr "Filtrovat podle"
-
 msgid "Filter parameters"
 msgstr "Parametry filtru"
 
@@ -2004,6 +2007,12 @@ msgstr "Jméno"
 
 msgid "Flag"
 msgstr "Příznak"
+
+msgid "Flag <strong><a href=\"{{ url }}\">{{ name }}</a></strong> created"
+msgstr "Příznak <strong><a href=\"{{ url }}\">{{ name }}</a></strong> byl vytvořen"
+
+msgid "Flag <strong><a href=\"{{ url }}\">{{ name }}</a></strong> modified"
+msgstr "Příznak <strong><a href=\"{{ url }}\">{{ name }}</a></strong> byl upraven"
 
 msgid "Flag <strong>{{ name }}</strong> deleted"
 msgstr "Příznak <strong>{{ name }}</strong> byl smazán"
@@ -2646,6 +2655,9 @@ msgstr "Nová skupina práv"
 
 msgid "New delivery address"
 msgstr "Nová doručovací adresa"
+
+msgid "New flag"
+msgstr "Nový příznak"
 
 msgid "New item"
 msgstr "Nová položka"
@@ -4117,6 +4129,9 @@ msgstr ""
 msgid "The product is no longer in the catalog. This was the catalog number at the time of the inquiry."
 msgstr "Produkt již není v katalogu. Toto bylo katalogové číslo v době vytvoření poptávky."
 
+msgid "The selected flag cannot be deleted."
+msgstr "Vybraný příznak nelze smazat."
+
 msgid "The stock quantities are set for the product variants separately."
 msgstr "Skladové zásoby se nastavují zvlášť pro jednotlivé varianty."
 
@@ -4143,6 +4158,9 @@ msgstr "Tuto měnu nelze smazat, je nastavena jako výchozí nebo je uložena u 
 
 msgid "This email was sent from the contact form located below the panel of categories."
 msgstr "Tento e-mail byl odeslán z kontaktního formuláře umístěného pod panelem kategorií."
+
+msgid "This flag can not be deleted as it is used in definition of"
+msgstr "Tento příznak nelze smazat, protože se používá v definici"
 
 msgid "This is the last customer user. It will remove all its related data. Do you really want to delete it ?"
 msgstr "Toto je poslední zákaznický účet. Odstraní všechna související data. Opravdu to chcete smazat?"
