+++ packages/framework/src/Resources/translations/validators.en.po
@@ -103,7 +103,7 @@ msgstr ""
 msgid "Flag \"{{ flagName }}\" is selected multiple times."
 msgstr ""
 
-msgid "Flag color in must be in valid hexadecimal code e.g. #3333ff"
+msgid "Flag color must be in valid hexadecimal code e.g. #3333ff"
 msgstr ""
 
 msgid "Flag name cannot be longer than {{ limit }} characters"
