+++ packages/framework/src/Resources/views/Admin/Content/Flag/deleteForbidden.html.twig
