+++ packages/framework/src/Resources/views/Admin/Content/Flag/detail.html.twig
