+++ packages/framework/src/Resources/views/Admin/Content/Flag/edit.html.twig
@@ -1,4 +1,4 @@
-{% extends 'Admin/Content/Flag/detail.html.twig' %}
+{% extends '@ShopsysFramework/Admin/Content/Flag/detail.html.twig' %}
 
 {% block title %}- {{ 'Editing flag'|trans }}{% endblock %}
 {% block h1 %}{{ 'Editing flag'|trans }}{% endblock %}
