+++ packages/framework/src/Resources/views/Admin/Content/Flag/list.html.twig
@@ -6,3 +6,12 @@
 {% block main_content %}
     {{ gridView.render() }}
 {% endblock %}
+
+{% block btn %}
+    <div class="wrap-bar">
+        <a href="{{ url('admin_flag_new') }}" class="btn btn--primary btn--plus wrap-bar__btn">
+            <i class="btn__icon">+</i>
+            {{ 'Create flag'|trans }}
+        </a>
+    </div>
+{% endblock %}
