+++ packages/framework/src/Resources/views/Admin/Content/Flag/listGrid.html.twig
@@ -9,11 +9,6 @@
     {{ parent() }}
 {% endblock %}
 
-{% block grid_inline_edit_add_button %}
-    {% set addEntity = 'Create flag'|trans %}
-    {{ parent() }}
-{% endblock %}
-
 {% block grid_value_cell_id_rgbColor %}
     {% if value is not empty %}
         <span class="in-pill" style="background-color: {{ value }};">{{ value }}</span>
