+++ packages/framework/src/Resources/views/Admin/Content/Flag/new.html.twig
@@ -1,4 +1,4 @@
-{% extends 'Admin/Content/Flag/detail.html.twig' %}
+{% extends '@ShopsysFramework/Admin/Content/Flag/detail.html.twig' %}
 
 {% block title %}- {{ 'New flag'|trans }}{% endblock %}
 {% block h1 %}{{ 'New flag'|trans }}{% endblock %}
