+++ project-base/app/src/Controller/Admin/SideMenuConfigurationSubscriber.php
@@ -79,10 +79,6 @@ class SideMenuConfigurationSubscriber implements EventSubscriberInterface
         $listMenu->removeChild('availabilities');
         $listMenu->getChild('units')->setLabel(t('Measurement units'));
 
-        $flagsMenu = $listMenu->getChild('flags');
-        $flagsMenu->addChild('flagNew', ['route' => 'admin_flag_new', 'label' => t('New flag'), 'display' => false]);
-        $flagsMenu->addChild('flagEdit', ['route' => 'admin_flag_edit', 'label' => t('Editing flag'), 'display' => false]);
-
         $storeMenu = $listMenu->addChild('stores', ['route' => 'admin_store_list', 'label' => t('Stores')]);
         $storeMenu->addChild('new_store', ['route' => 'admin_store_new', 'display' => false, 'label' => t('New store')]);
         $storeMenu->addChild('edit_store', ['route' => 'admin_store_edit', 'display' => false, 'label' => t('Edit store')]);
