+++ project-base/app/src/Model/Product/Filter/Elasticsearch/ProductFilterConfigFactory.php
@@ -12,14 +12,12 @@ use Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterConfigFactory as B
  * @property \App\Model\Product\Filter\ParameterFilterChoiceRepository $parameterFilterChoiceRepository
  * @property \App\Model\Product\Filter\FlagFilterChoiceRepository $flagFilterChoiceRepository
  * @property \App\Model\Product\Filter\PriceRangeRepository $priceRangeRepository
- * @property \App\Model\Product\Filter\BrandFilterChoiceRepository $brandFilterChoiceRepository
  * @property \App\Model\Customer\User\CurrentCustomerUser $currentCustomerUser
- * @property \App\Model\Product\Flag\FlagFacade $flagFacade
  * @method \Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterConfig createForCategory(string $locale, \App\Model\Category\Category $category)
  * @method \Shopsys\FrameworkBundle\Model\Product\Filter\ParameterFilterChoice[] getSortedParameterFilterChoicesForCategory(\Shopsys\FrameworkBundle\Model\Product\Filter\ParameterFilterChoice[] $aggregatedParameterFilterChoices, \App\Model\Category\Category $category)
  * @method \Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterConfig create(\Shopsys\FrameworkBundle\Model\Product\Filter\ParameterFilterChoice[] $parameterChoices, \App\Model\Product\Flag\Flag[] $flagChoices, \App\Model\Product\Brand\Brand[] $brandChoices, \Shopsys\FrameworkBundle\Model\Product\Filter\PriceRange $priceRange)
  * @property \App\Model\Product\Parameter\ParameterFacade $parameterFacade
- * @method __construct(\App\Model\Product\Filter\ParameterFilterChoiceRepository $parameterFilterChoiceRepository, \App\Model\Product\Filter\FlagFilterChoiceRepository $flagFilterChoiceRepository, \App\Model\Customer\User\CurrentCustomerUser $currentCustomerUser, \Shopsys\FrameworkBundle\Model\Product\Filter\BrandFilterChoiceRepository $brandFilterChoiceRepository, \App\Model\Product\Filter\PriceRangeRepository $priceRangeRepository, \Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterElasticFacade $productFilterElasticFacade, \App\Model\Product\Parameter\ParameterFacade $parameterFacade, \App\Model\Product\Flag\FlagFacade $flagFacade, \Shopsys\FrameworkBundle\Model\Product\Brand\BrandFacade $brandFacade)
+ * @method __construct(\App\Model\Product\Filter\ParameterFilterChoiceRepository $parameterFilterChoiceRepository, \App\Model\Product\Filter\FlagFilterChoiceRepository $flagFilterChoiceRepository, \App\Model\Customer\User\CurrentCustomerUser $currentCustomerUser, \Shopsys\FrameworkBundle\Model\Product\Filter\BrandFilterChoiceRepository $brandFilterChoiceRepository, \App\Model\Product\Filter\PriceRangeRepository $priceRangeRepository, \Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterElasticFacade $productFilterElasticFacade, \App\Model\Product\Parameter\ParameterFacade $parameterFacade, \Shopsys\FrameworkBundle\Model\Product\Flag\FlagFacade $flagFacade, \Shopsys\FrameworkBundle\Model\Product\Brand\BrandFacade $brandFacade)
  * @method \Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterConfig createForFlag(\App\Model\Product\Flag\Flag $flag, string $locale)
  */
 class ProductFilterConfigFactory extends BaseProductFilterConfigFactory
