+++ project-base/app/src/Model/Product/Flag/FlagData.php
@@ -4,20 +4,8 @@ declare(strict_types=1);
 
 namespace App\Model\Product\Flag;
 
-use Shopsys\FrameworkBundle\Component\Router\FriendlyUrl\UrlListData;
 use Shopsys\FrameworkBundle\Model\Product\Flag\FlagData as BaseFlagData;
 
 class FlagData extends BaseFlagData
 {
-    /**
-     * @var \Shopsys\FrameworkBundle\Component\Router\FriendlyUrl\UrlListData
-     */
-    public $urls;
-
-    public function __construct()
-    {
-        parent::__construct();
-
-        $this->urls = new UrlListData();
-    }
 }
