--- project-base/app/templates/Admin/Content/Flag/list.html.twig
@@ -1,10 +1,0 @@
-{% extends '@!ShopsysFramework/Admin/Content/Flag/list.html.twig' %}
-
-{% block btn %}
-    <div class="wrap-bar">
-        <a href="{{ url('admin_flag_new') }}" class="btn btn--primary btn--plus wrap-bar__btn">
-            <i class="btn__icon">+</i>
-            {{ 'Create flag'|trans }}
-        </a>
-    </div>
-{% endblock %}
