+++ project-base/app/translations/messages.cs.po
@@ -40,9 +40,6 @@ msgstr "Banner"
 msgid "Banners"
 msgstr "Bannery"
 
-msgid "Basic information"
-msgstr "Základní informace"
-
 msgid "Basic objects"
 msgstr "Basic"
 
@@ -51,9 +48,6 @@ msgstr "Přehled značek"
 
 msgid "CMS"
 msgstr "CMS"
-
-msgid "Cancel"
-msgstr "Zrušit"
 
 msgid "Catalog numbers"
 msgstr "Katalogová čísla"
@@ -66,9 +60,6 @@ msgstr "Vymazat cache s graphql dotazy na FE"
 
 msgid "Close"
 msgstr "Zavřít"
-
-msgid "Color"
-msgstr "Barva"
 
 msgid "Column 1"
 msgstr "Sloupec 1"
@@ -85,17 +76,11 @@ msgstr "Kontakt"
 msgid "Content"
 msgstr "Obsah"
 
-msgid "Create flag"
-msgstr "Vytvořit příznak"
-
 msgid "Create new banner"
 msgstr "Vytvořit nový banner"
 
 msgid "Create notification bar"
 msgstr "Vytvořit notifikační lištu"
-
-msgid "Creating a record requires all domains to be enabled as domain-specific fields cannot be empty. If you want to proceed, select all domains in the Domain filter in the header first."
-msgstr "Vytvoření záznamu vyžaduje, aby byly povoleny všechny domény, protože pole specifická pro doménu nemohou být prázdná. Pokud chcete pokračovat, nejprve vyberte všechny domény ve filtru domény v záhlaví."
 
 msgid "Default"
 msgstr "Základní"
@@ -106,17 +91,11 @@ msgstr "Popis"
 msgid "Disable two factor authentication"
 msgstr "Vypnout dvoufaktorové ověřování"
 
-msgid "Display"
-msgstr "Zobrazit"
-
 msgid "Display date FROM"
 msgstr "Datum zobrazení OD"
 
 msgid "Display date TO"
 msgstr "Datum zobrazení DO"
-
-msgid "Do you really want to remove this flag?"
-msgstr "Opravdu chcete odstranit tento příznak?"
 
 msgid "Do you really want to remove this language constant translation?"
 msgstr "Opravdu chcete odstranit tento překlad jazykové konstanty?"
@@ -136,12 +115,6 @@ msgstr "Upravit prodejnu"
 msgid "Editing banner"
 msgstr "Upravit banner"
 
-msgid "Editing flag"
-msgstr "Editace příznaku"
-
-msgid "Editing flag - {{ name }}"
-msgstr "Editace příznaku - {{ name }}"
-
 msgid "Editing notification bar"
 msgstr "Editace notifikační lišty"
 
@@ -154,26 +127,11 @@ msgstr "Zapnout dvoufaktorové ověřování pomocí Google Autentikátoru"
 msgid "Enable two factor authentication by email"
 msgstr "Zapnout dvoufaktorové ověřování pomocí e-mailu"
 
-msgid "Extended SEO categories"
-msgstr "Rozšířené SEO kategorie"
-
 msgid "File"
 msgstr "Soubor"
 
 msgid "File not found in main filesystem. Please try again later."
 msgstr "Soubor se nepodařilo nalézt na hlavním uložišti. Zkuste to, prosím znovu později."
-
-msgid "Flag"
-msgstr "Příznak"
-
-msgid "Flag <strong><a href=\"{{ url }}\">{{ name }}</a></strong> created"
-msgstr "Příznak <strong><a href=\"{{ url }}\">{{ name }}</a></strong> byl vytvořen"
-
-msgid "Flag <strong><a href=\"{{ url }}\">{{ name }}</a></strong> modified"
-msgstr "Příznak <strong><a href=\"{{ url }}\">{{ name }}</a></strong> byl upraven"
-
-msgid "Flag <strong>{{ name }}</strong> deleted"
-msgstr "Příznak <strong>{{ name }}</strong> byl smazán"
 
 msgid "Forgotten password"
 msgstr "Zapomenuté heslo"
@@ -307,9 +265,6 @@ msgstr "Nový banner"
 msgid "New closed day"
 msgstr "Nový svátek / interní den"
 
-msgid "New flag"
-msgstr "Nový příznak"
-
 msgid "New notification bar"
 msgstr "Nová notifikační lišta"
 
@@ -394,9 +349,6 @@ msgstr "Nastavení zásad ochrany osobních údajů"
 msgid "Product image"
 msgstr "Obrázek produktu"
 
-msgid "Promo code"
-msgstr "Slevový kupón"
-
 msgid "Promoted categories"
 msgstr "Promované kategorie"
 
@@ -436,14 +388,8 @@ msgstr "Vyhledat"
 msgid "Secondary"
 msgstr "Sekundární"
 
-msgid "Selected flag doesn't exist."
-msgstr "Vybraný příznak neexistuje."
-
 msgid "Selected notification bar does not exist"
 msgstr "Vybraná notifikační lišta neexistuje"
-
-msgid "Seo"
-msgstr "Seo"
 
 msgid "Settings"
 msgstr "Nastavení"
@@ -478,20 +424,11 @@ msgstr "Text"
 msgid "Text with image"
 msgstr "Text s obrázkem"
 
-msgid "The selected flag cannot be deleted."
-msgstr "Vybraný příznak nelze smazat."
-
-msgid "This flag can not be deleted as it is used in definition of"
-msgstr "Tento příznak nelze smazat, protože se používá v definici"
-
 msgid "Title"
 msgstr "Title"
 
 msgid "Type of image"
 msgstr "Typ obrázku"
-
-msgid "URL addresses"
-msgstr "URL adresy"
 
 msgid "Unable to load list of language constants"
 msgstr "Nepodařilo se načíst přehled jazykových konstant"
