+++ project-base/app/translations/validators.cs.po
@@ -13,9 +13,6 @@ msgstr "E-mail nesmí být delší než {{ limit }} znaků"
 msgid "Flag color must be in valid hexadecimal code e.g. #3333ff"
 msgstr "Barva příznaku musí mít správný formát hexadecimálního kódu, např.: #3333ff"
 
-msgid "Flag name cannot be longer than {{ limit }} characters"
-msgstr "Název příznaku nemůže být delší než {{ limit }} znaků"
-
 msgid "Image can be only in JPG or PNG format"
 msgstr "Obrázek může být pouze ve formátu JPG nebo PNG"
 
@@ -39,9 +36,6 @@ msgstr "Vyplňte prosím e-mail"
 
 msgid "Please enter flag color"
 msgstr "Vyplňte prosím barvu příznaku"
-
-msgid "Please enter flag name in all languages"
-msgstr "Prosím zadejte název příznaku ve všech jazycích"
 
 msgid "Please enter new password"
 msgstr "Vyplňte prosím nové heslo"
