+++ project-base/app/translations/validators.en.po
@@ -13,9 +13,6 @@ msgstr ""
 msgid "Flag color must be in valid hexadecimal code e.g. #3333ff"
 msgstr ""
 
-msgid "Flag name cannot be longer than {{ limit }} characters"
-msgstr ""
-
 msgid "Image can be only in JPG or PNG format"
 msgstr ""
 
@@ -38,9 +35,6 @@ msgid "Please enter email"
 msgstr ""
 
 msgid "Please enter flag color"
-msgstr ""
-
-msgid "Please enter flag name in all languages"
 msgstr ""
 
 msgid "Please enter new password"
